ai_workload_analyzer/pyproject.toml
[project]
name = "ai-workload-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncio-mqtt>=0.16.2",
    "chainlit>=2.6.0",
    "databricks-agents>=1.1.0",
    "databricks-langchain>=0.6.0",
    "langchain>=0.3.26",
    "langchain-core>=0.3.68",
    "numpy>=1.26.4",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
